src/requirements.txt
fastapi[standard]==0.116.1
uvicorn==0.35.0
python-multipart==0.0.19
python-dotenv==1.0.0
pydantic-settings==2.10.1
aiofiles==24.1.0
langchain==0.3.27
PyMuPDF==1.26.4
langchain-community==0.3.29
